+++ apps/web/src/components/TopicSelector/TopicSelector.js
@@ -35,6 +35,10 @@ const inputStyles = {
       border: '1px solid hsl(var(--border))'
     }
   }),
+  input: styles => ({
+    ...styles,
+    color: 'hsl(var(--text-foreground))'
+  }),
   multiValue: styles => ({
     ...styles,
     backgroundColor: 'hsl(var(--selected))',
@@ -125,8 +129,8 @@ function TopicSelector (props) {
 
     if (selected.length >= MAX_TOPICS || isEmpty(input)) return []
 
-    const response = await dispatch(findTopics({ autocomplete: input, groupSlug: slug }))
-    const topicResults = response.payload.getData().items.map(get('topic'))
+    const response = await dispatch(findTopics({ autocomplete: input, groupSlug: slug, includeCounts: true }))
+    const topicResults = response.payload.getData().items
     const sortedTopicResults = orderBy(
       [t => t.name === input ? -1 : 1, 'followersTotal', 'postsTotal'],
       ['asc', 'desc', 'desc'],