+++ apps/web/src/routes/GroupSettings/RelatedGroupsTab/RelatedGroupsTab.js
@@ -128,7 +128,7 @@ function RelatedGroupsTab () {
                 <GroupCard
                   group={p}
                   key={p.id}
-                  actionMenu={<Dropdown toggleChildren={<Icon name='More' />} items={relationshipDropdownItems(p, group, GROUP_RELATIONSHIP_TYPE.ChildToParent)} className='right-0 left-auto' />}
+                  actionMenu={<Dropdown id='related-groups-parent-dropdown' toggleChildren={<Icon name='More' />} items={relationshipDropdownItems(p, group, GROUP_RELATIONSHIP_TYPE.ChildToParent)} className='right-0 left-auto' />}
                 />
               ))}
             </div>
@@ -210,7 +210,7 @@ function RelatedGroupsTab () {
                 <GroupCard
                   group={c}
                   key={c.id}
-                  actionMenu={<Dropdown toggleChildren={<Icon name='More' />} items={relationshipDropdownItems(group, c, GROUP_RELATIONSHIP_TYPE.ParentToChild)} className='right-0 left-auto' />}
+                  actionMenu={<Dropdown id='related-groups-child-dropdown' toggleChildren={<Icon name='More' />} items={relationshipDropdownItems(group, c, GROUP_RELATIONSHIP_TYPE.ParentToChild)} className='right-0 left-auto' />}
                 />
               ))}
             </div>