+++ apps/web/src/routes/Groups/Groups.jsx
@@ -1,31 +1,18 @@
-import { Link } from 'react-router-dom'
 import React, { useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
 import { Helmet } from 'react-helmet'
 import { useSelector } from 'react-redux'
 import { createSelector } from 'reselect'
-import { bgImageStyle, cn } from 'util/index'
-import Icon from 'components/Icon'
-import RoundImage from 'components/RoundImage'
+import { cn } from 'util/index'
 import GroupNetworkMap from 'components/GroupNetworkMap'
 import { useViewHeader } from 'contexts/ViewHeaderContext'
 import { useGetJoinRequests } from 'hooks/useGetJoinRequests'
 import useRouteParams from 'hooks/useRouteParams'
-import {
-  DEFAULT_BANNER,
-  DEFAULT_AVATAR,
-  accessibilityDescription,
-  accessibilityIcon,
-  accessibilityString,
-  visibilityDescription,
-  visibilityIcon,
-  visibilityString
-} from 'store/models/Group'
 import getGroupForSlug from 'store/selectors/getGroupForSlug'
 import { getChildGroups, getParentGroups } from 'store/selectors/getGroupRelationships'
 import getMyMemberships from 'store/selectors/getMyMemberships'
-import { groupUrl, groupDetailUrl } from 'util/navigation'
 import { mapNodesAndLinks } from 'util/networkMap'
+import GroupCard from 'components/GroupCard'
 
 import classes from './Groups.module.scss'
 
@@ -127,48 +114,4 @@ export function GroupsList ({ groups, routeParams }) {
   )
 }
 
-export function GroupCard ({ group, routeParams }) {
-  const { t } = useTranslation()
-  return (
-    <Link to={group.memberStatus === 'member' ? groupUrl(group.slug) : groupDetailUrl(group.slug, routeParams)}>
-      <div className='flex relative rounded-lg p-4 bg-black shadow-xl hover:scale-102 transition-all duration-300'>
-        <div className='flex gap-2 relative z-10 w-full justify-between'>
-          <div className='flex flex-row gap-2'>
-            <RoundImage url={group.avatarUrl || DEFAULT_AVATAR} size='50px' square className='shadow-xl' />
-            <div className={cn('flex flex-row gap-2', classes.groupDetails)}>
-              <div className='flex flex-col gap-0'>
-                <span className={classes.groupName}>{group.name}</span>
-                {group.memberCount ? <span className='text-xs text-white/80'>{group.memberCount} {t('Members')}</span> : ''}
-              </div>
-            </div>
-          </div>
-          <div className={cn(classes.groupStats)}>
-            <div className={cn('flex flex-row gap-2 items-center h-full', classes.membershipStatus)}>
-              <div className={classes.groupPrivacy}>
-                <Icon name={visibilityIcon(group.visibility)} className={classes.privacyIcon} />
-                <div className={classes.privacyTooltip}>
-                  <div><strong>{t(visibilityString(group.visibility))}</strong> - {t(visibilityDescription(group.visibility))}</div>
-                </div>
-              </div>
-              <div className={classes.groupPrivacy}>
-                <Icon name={accessibilityIcon(group.accessibility)} className={classes.privacyIcon} />
-                <div className={classes.privacyTooltip}>
-                  <div><strong>{t(accessibilityString(group.accessibility))}</strong> - {t(accessibilityDescription(group.accessibility))}</div>
-                </div>
-              </div>
-              {
-                group.memberStatus === 'member'
-                  ? <div className={classes.statusTag}><Icon name='Complete' className={classes.memberComplete} /> <b>{t('Member')}</b></div>
-                  : group.memberStatus === 'requested'
-                    ? <div className={classes.statusTag}><b>{t('Membership Requested')}</b></div>
-                    : <div className='focus:text-foreground relative text-base border-2 border-foreground/20 hover:border-foreground/100 hover:text-foreground rounded-md p-2 bg-background text-foreground w-full transition-all scale-100 hover:scale-102 opacity-85 hover:opacity-100 flex items-center'><Icon name='CirclePlus' className={classes.joinGroup} /> <b>{t('Join')}</b></div>
-              }
-            </div>
-          </div>
-        </div>
-        <div style={bgImageStyle(group.bannerUrl || DEFAULT_BANNER)} className='w-full h-full bg-cover bg-center rounded-lg absolute top-0 left-0 opacity-70' />
-      </div>
-    </Link>
-  )
-}
 export default Groups